Extract startup diagnostics core microcrate
