+++ Cargo.toml
@@ -82,6 +82,7 @@ members = [
   "crates/bitnet-testing-policy-runtime-core",
   "crates/bitnet-testing-policy-state-core",
   "crates/bitnet-startup-contract-core",
+  "crates/bitnet-startup-contract-diagnostics-core",
   "crates/bitnet-startup-contract-diagnostics",
   "crates/bitnet-startup-contract-guard",
   "crates/bitnet-test-support",
